Move background map functions to background.py
