+++ gammapy/cube/__init__.py
@@ -5,6 +5,7 @@ Sky cubes (3-dimensional: energy, lon, lat).
 from .models import *
 from .counts import *
 from .exposure import *
+from .background import *
 from .fit import *
 from .new import *
 from .psf_kernel import *
